router/babydiary/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anthropic==0.31.1
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arxiv==2.1.3
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
boto3==1.34.144
botocore==1.34.144
build==1.2.1
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
cohere==5.5.8
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
dataclasses-json==0.6.7
datasets==2.14.4
debugpy==1.8.2
decorator==5.1.1
deepl==1.18.0
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
distro==1.9.0
dnspython==2.6.1
EbookLib==0.18
effdet==0.4.1
elastic-transport==8.13.1
elasticsearch==8.14.0
email_validator==2.2.0
emoji==2.12.1
environs==9.5.0
et-xmlfile==1.1.0
executing==2.0.1
faiss-cpu==1.8.0.post1
fastapi==0.111.1
fastapi-cli==0.0.4
fastavro==1.9.5
fastjsonschema==2.20.0
feedparser==6.0.11
filelock==3.15.4
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==1.34.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
google_search_results==2.4.2
googleapis-common-protos==1.63.2
gpt4all==2.7.0
greenlet==3.0.3
grpc-gateway-protoc-gen-openapiv2==0.1.0
grpcio==1.63.0
grpcio-status==1.48.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
iopath==0.1.10
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
JPype1==1.5.0
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
kiwipiepy==0.17.1
kiwipiepy_model==0.17.0
kiwisolver==1.4.5
konlpy==0.6.0
kubernetes==30.1.0
langchain==0.2.7
langchain-anthropic==0.1.20
langchain-chroma==0.1.2
langchain-cohere==0.1.9
langchain-community==0.2.7
langchain-core==0.2.18
langchain-elasticsearch==0.2.2
langchain-experimental==0.0.61
langchain-google-genai==1.0.7
langchain-huggingface==0.0.3
langchain-milvus==0.1.1
langchain-openai==0.1.16
langchain-teddynote==0.0.11
langchain-text-splitters==0.2.2
langchain-upstage==0.1.7
langchainhub==0.1.20
langdetect==1.0.9
langgraph==0.0.69
langsmith==0.1.85
lark==1.1.9
layoutparser==0.3.4
lxml==5.2.2
lz4==4.3.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.5
multiprocess==0.70.15
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.14.1
onnxruntime==1.18.1
open_clip_torch==2.26.1
openai==1.35.13
opencv-python==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
orjson==3.10.6
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parameterized==0.9.0
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.2
pillow==10.4.0
pinecone-client==3.2.2
pinecone-text==0.9.0
platformdirs==4.2.2
portalocker==2.10.1
posthog==3.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==3.20.3
psutil==6.0.0
pulsar-client==3.5.0
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.8
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pymilvus==2.4.4
PyMuPDF==1.24.7
pymupdf4llm==0.0.9
PyMuPDFb==1.24.6
pypandoc==1.13
pyparsing==3.1.2
pypdf==4.3.0
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.4.1
pysbd==0.3.4
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.9
python-pptx==0.6.21
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
ragas==0.1.10
rank-bm25==0.2.2
rapidfuzz==3.9.4
redis==5.0.7
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.19.0
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
Send2Trash==1.8.3
sentence-transformers==3.0.1
sgmllib3k==1.0.0
shellingham==1.5.4
simsimd==4.4.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
stack-data==0.6.3
starlette==0.37.2
streamlit==1.36.0
sympy==1.13.0
tabulate==0.9.0
tenacity==8.3.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==1.0.7
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.2.2
torchvision==0.17.2
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.42.4
typer==0.12.3
types-python-dateutil==2.9.0.20240316
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
unstructured==0.10.19
unstructured-inference==0.6.6
unstructured.pytesseract==0.3.12
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.1
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
wget==3.2
widgetsnbextension==4.0.11
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.2.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
